requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
pyperclip==1.8.2
google-api-python-client==2.120.0
python-dotenv==1.0.1
google-generativeai==0.3.2
markdown2==2.4.12
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0